requirements.txt
absl-py==2.0.0
astunparse==1.6.3
backports.tarfile==1.0.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
danila-lib==1.9.5
docutils==0.21.1
easyocr==1.7.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
Jinja2==3.1.2
keras==3.3.3
keyring==25.1.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nh3==0.2.17
ninja==1.11.1.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.4
pillow==10.3.0
pkginfo==1.10.0
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
Pygments==2.17.2
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readme_renderer==43.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.12.9
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
twine==5.0.0
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.2.2
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0